% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.ppbr_fit}
\alias{plot.ppbr_fit}
\title{Plot a PPBR analysis}
\usage{
\method{plot}{ppbr_fit}(x, type = c("target", "states", "brackets", "safety", "all"), ...)
}
\arguments{
\item{x}{A `ppbr_fit` object.}

\item{type}{One of `"target"`, `"states"`, `"brackets"`, `"safety"`, or
`"all"`.}

\item{...}{Additional graphical arguments.}
}
\description{
Plot a PPBR analysis
}
