% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.ppbr_simulation}
\alias{plot.ppbr_simulation}
\title{Plot PPBR simulation results}
\usage{
\method{plot}{ppbr_simulation}(x, type = c("selection", "allocation"), ...)
}
\arguments{
\item{x}{A `ppbr_simulation` object.}

\item{type}{`"selection"` or `"allocation"`.}

\item{...}{Additional graphical arguments.}
}
\description{
Plot PPBR simulation results
}
