% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppbr-package.R
\docType{package}
\name{ppbr-package}
\alias{ppbr-package}
\title{ppbr: Posterior Predictive Bracketing and Resolution}
\description{
A compact implementation of PPBR for phase I dose-finding trials with
ordered doses and binary toxicity outcomes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/haohaostats/PPBR}
  \item Report bugs at \url{https://github.com/haohaostats/PPBR/issues}
}

}
\author{
\strong{Maintainer}: PPBR Development Team \email{haocstats@gmail.com}

}
\keyword{internal}
