% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{ppbr}
\alias{ppbr}
\title{Fit PPBR to the current dose-level data}
\usage{
ppbr(design, n, dlt, current_dose = NULL)
}
\arguments{
\item{design}{A `ppbr_design` object.}

\item{n}{Number of patients treated at each dose.}

\item{dlt}{Number of dose-limiting toxicities at each dose.}

\item{current_dose}{Current dose label. If omitted, the last administered
dose in the ordered grid is used; an empty trial uses the starting dose.}
}
\value{
An object of class `ppbr_fit`.
}
\description{
Fit PPBR to the current dose-level data
}
