% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/next.R
\name{ppbr_next}
\alias{ppbr_next}
\title{Obtain the next PPBR action}
\usage{
ppbr_next(fit)
}
\arguments{
\item{fit}{A fitted `ppbr_fit` object.}
}
\value{
An object of class `ppbr_decision`.
}
\description{
Obtain the next PPBR action
}
